refactor code, handle process dead

+++ PokemonReviewApp/Models/InputAudioConverter.cs
@@ -4,6 +4,6 @@
     {
         public string InputUrl { get; set; } = default!;
         public bool RecordToMp3 { get; set; }
-        public long RecordTimeInSec { get; set; }
+        public int RecordTimeInSec { get; set; }
     }
 }
